requirements.txt
PyJWT==2.3.0
fastapi==0.70.0
cryptography==36.0.0
uvicorn==0.15.0
